+++ bot/handlers/main_message_handlers.py
@@ -2,12 +2,12 @@ from aiogram import F, Router, Dispatcher
 from aiogram.types import Message
 
 from api.app.user.schemas import UserCreate
-from bot.common.crud.api.user import register_user, login_user, get_user, update_user
+from bot.common.crud.api.user import register_user, login_user, get_user
 from bot.common.crud.bot.user_info import create_user_info, get_user_info, update_user_info
 from bot.common.utils import get_text
 from bot.config import language_buttons, buttons
-from bot.handlers.button_handlers import button_handlers, handle_profile
-from bot.handlers.display_handlers import get_contact_keyboard, get_reply_keyboard
+from bot.handlers.keyboard_handlers import get_contact_keyboard, get_reply_keyboard
+from bot.handlers.reply_buttons_handlers import button_handlers
 
 router = Router()
 
@@ -63,48 +63,6 @@ async def handle_contact(message: Message):
     )
 
 
-@router.message(
-    F.text.contains("Прізвище:") |
-    F.text.contains("Ім'я:") |
-    F.text.contains("First name:") |
-    F.text.contains("Last name:")
-)
-async def process_profile_update(message: Message):
-    language_code = get_user_info(message.from_user.id).language_code
-
-    lines = message.text.strip().split("\n")
-
-    field_mappings = {
-        "First name:": "first_name",
-        "Last name:": "last_name",
-
-        "Ім'я:": "first_name",
-        "Прізвище:": "last_name",
-    }
-
-    try:
-        raw_dict = dict(item.split(': ', 1) for item in lines if ': ' in item)
-    except ValueError:
-        await message.answer(
-            get_text("invalid_format", language_code)
-        )
-        return
-
-    data_for_update = {field_mappings.get(k + ":", k): v for k, v in raw_dict.items() if v != ''}
-
-    if not data_for_update:
-        await message.answer(get_text("invalid_format", language_code))
-        return
-
-    await update_user(message.from_user.id, **data_for_update)
-
-    await message.answer(
-        get_text('profile_updated', language_code)
-    )
-
-    await handle_profile(message, language_code)
-
-
 @router.message()
 async def handle_buttons(message: Message):
     telegram_id = message.from_user.id
@@ -123,5 +81,5 @@ async def handle_buttons(message: Message):
         await message.answer(get_text("use_buttons", language_code))
 
 
-def register_message_handlers(dispatcher: Dispatcher):
+def register_main_message_handlers(dispatcher: Dispatcher):
     dispatcher.include_router(router)
