+++ bot/main.py
@@ -6,9 +6,10 @@ from aiogram.fsm.storage.memory import MemoryStorage
 
 from bot.common.database import create_db_and_tables
 from bot.config import TG_TOKEN
-from bot.handlers.edit_handlers import register_edit_handlers
 from bot.handlers.command_handlers import register_command_handlers
-from bot.handlers.message_handlers import register_message_handlers
+from bot.handlers.edit_handlers import register_edit_handlers
+from bot.handlers.main_message_handlers import register_main_message_handlers
+from bot.handlers.profile_handlers.edit_handler import register_profile_edit_handlers
 
 
 async def main():
@@ -17,9 +18,10 @@ async def main():
     bot = Bot(token=TG_TOKEN)
     dispatcher = Dispatcher(storage=MemoryStorage())
 
+    register_profile_edit_handlers(dispatcher)
     register_command_handlers(dispatcher)
-    register_message_handlers(dispatcher)
     register_edit_handlers(dispatcher)
+    register_main_message_handlers(dispatcher)
 
     await dispatcher.start_polling(bot)
